bugs fixed, maybe not
I forgot what bugs I had to fix yesterday, but I fixed the ones I found.
BUG!!!: If you log out and then log back in, messages waiting for delivery will not be delivered. 

client: a lot of code is spread across functions for convenience, now you can get messages that were sent during the absence of the network.

server: now the server does not break when you try to send a message to a client that is not online.

database: Added new table for messages waiting for delivery

constants: a New constant, of course

common_functions_and_data_structures: slightly corrected the function, maybe even a couple of bugs fixed, and maybe a few added.