new features (from last commit, still doesnt work)
now I do not have to convert a string into bytes myself, it is put into a function that prepares the string for sending. Fixed receiving messages when in some cases the message itself could be received together with the message data. But if we managed to get data about the next message, then GG, has not yet figured out how to fix.