the database is complete (or almost complete)
All necessary methods of data storage and processing in the database are written. The client has written a request to receive data from the server, and the server is written to send this data. But receiving data in the client is not written yet. But the processing of messages stored by the client seems to work. Not written the establishment of a secret chat using the server for p2p similar. All that is written has never been started, so most likely nothing works.