eto pobeda (victory!)
Any types of messages are transmitted between users, and are stored in the database, and restored from it when required. This is the main part of the work with the messenger is over. In the future, I will need to come up with a verification procedure in the p2p connection, and put the code in order (there are still a lot of bugs). But first I need to write a GUI because i will have to show it soon .