the greatest GUI this world has ever seen
On the basis of gui written gui2, which unlike its predecessor performs its basic functions. All the basic functions of the GUI mostly working. The connection of GUI with the client and the client with GUI is established. At this moment I consider writing the GUI is over. In the future, of course, you can fix all the errors and make many necessary improvements, but now it is not a priority .